+++ src/libslic3r/CMakeLists.txt
@@ -221,8 +221,10 @@ set(SLIC3R_SOURCES
     JumpPointSearch.cpp
     JumpPointSearch.hpp
     KDTreeIndirect.hpp
+    IndexRange.hpp
     Layer.cpp
     Layer.hpp
+    LayerRegion.hpp
     LayerRegion.cpp
     libslic3r.h
     "${CMAKE_CURRENT_BINARY_DIR}/libslic3r_version.h"
