+++ tests/libslic3r/CMakeLists.txt
@@ -42,6 +42,7 @@ add_executable(${_TEST_NAME}_tests
     test_anyptr.cpp
     test_jump_point_search.cpp
     test_support_spots_generator.cpp
+    test_layer_region.cpp
     ../data/prusaparts.cpp
     ../data/prusaparts.hpp
      test_static_map.cpp
